python/requirements.txt
# Signal Scanner Python Dependencies

# Data fetching
yfinance>=0.2.36
pycoingecko>=3.1.0
pandas>=2.0.0

# Database
asyncpg>=0.29.0

# AI validation
google-genai>=1.0.0

# Telegram notifications
python-telegram-bot>=20.7

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Utilities
python-dotenv>=1.0.0
